+++ mstrio/utils/enums.py
@@ -1,4 +1,5 @@
 from enum import auto
+
 from mstrio.utils.enum_helper import AutoName